Move injection_points.h to pg_lake_engine and use it in client.c
The header was in pg_lake_table/include/, which pg_lake_engine cannot
reach (it sits lower in the dependency graph).  Moving it to
pg_lake_engine/include/pg_lake/util/ makes it accessible to both:
pg_lake_engine includes its own include/ tree, and pg_lake_table
already has -I../pg_lake_engine/include on its compiler flags.

client.c drops its hand-rolled shim and uses INJECTION_POINT_COMPAT
from the shared header instead.
